* Defined default dial timeout (5 seconds)
